general/requirements.txt
requests
discord.py
flask
colorama 
asyncio

